Move and update the streamer manager class
